Consulta projetos legislativos no site da Assembleia Legislativa do Estado do Paraná

Esta função realiza uma consulta aos projetos legislativos no site da Assembleia Legislativa do Estado do Paraná, com base nos parâmetros fornecidos. Os parâmetros podem incluir o tipo de proposição, autores, número, ano, conclusão, assuntos, chave e sumula.

Usage

alepr_projeto(
  prop_tipo = "",
  prop_autores = "",
  prop_numero = "",
  prop_ano = "",
  prop_conclusao = "",
  prop_assuntos = "",
  prop_chave = "",
  prop_sumula = "",
  .reportar = TRUE
)

Argumentos

prop_tipo

Tipo da proposição. Pode ser vazio (todos os tipos) ou um número de 1 a 11, onde:

1: Lei

2: Lei Complementar

3: Resolução

4: Decreto Legislativo

5: Emenda Constitucional

6: Indicação Legislativa

7: Ato

10: Ofício

11: Documento de Designação

12: Portaria

prop_autores

Autores da proposição.

prop_numero

Número da proposição.

prop_ano

Ano da proposição.

prop_conclusao

Status de conclusão da proposição. Pode ser vazio (todos os tipos) ou um número de 0 a 5, onde:

0: Sancionada

1: Promulgada

2: Sancionada/Republicada

3: Promulgada/Republicada

4: Revogada

5: Aprovada

prop_assuntos

Assuntos relacionados à proposição.

prop_chave

Chave da proposição.

prop_sumula

Sumula da proposição.

.reportar

Valor lógico indicando se deve reportar a ação realizada. Padrão é TRUE.

Valor

Retorna uma tabela com os projetos legislativos correspondentes aos parâmetros fornecidos. Se nenhum projeto for encontrado, retorna NULL.

Veja também

Exemplos

alepr_projeto(prop_tipo = 1, prop_numero = 21890, prop_ano = 2024)
#> Extrai do site da Assembleia Legislativa do Estado do Paraná as informações sobre a proposição que deu origem a norma nº 21890 de 2024

alepr_projeto(prop_tipo = 4, prop_numero = 29, prop_ano = 2020)
#> Extrai do site da Assembleia Legislativa do Estado do Paraná as informações sobre a proposição que deu origem a norma nº 29 de 2020