Extrai informações sobre as decisões de um processo judicial no STF

Esta função extrai informações sobre as decisões de um processo judicial no Supremo Tribunal Federal (STF) utilizando a API do portal STF. Ela recebe como entrada a classe do processo e o número do processo e retorna um data.frame contendo informações sobre as decisões do processo.

Usage

stf_decisoes(classe, processo)

Argumentos

classe

A classe do processo, em maiúsculas. Por exemplo, "RE", "HC", "ADI", etc.

processo

O número do processo ou uma lista de números de processo.

Valor

Um data.frame contendo informações sobre as decisões do processo, incluindo as seguintes colunas:

Classe

A classe do processo.

Data

A data da decisão.

Nome

O nome da decisão.

Julgador

O julgador responsável pela decisão.

Decisao

O conteúdo da decisão.

Detalhes

Para mais detalhes, acesse: vignette("familia_stf", package = "decJ")

Veja também

Exemplos

if (FALSE) {
stf_decisoes("ADPF", "800")
stf_decisoes("ADPF", c("800", "900"))
}