Baixa jurisprudência do Supremo Tribunal Federal (STF)

Esta função baixa jurisprudência do Supremo Tribunal Federal (STF) utilizando a API do portal de jurisprudência STF. Você pode pesquisar por palavras-chave ou por classe de processo e especificar a quantidade de documentos a serem baixados. Os documentos são salvos como arquivos PDF.

Usage

stf_jurisprudencia_download(
  busca = NULL,
  classe = NULL,
  base = c("acordaos", "decisoes"),
  quantidade = 25,
  arquivo = "."
)

Argumentos

busca

Uma string contendo palavras-chave para a busca de jurisprudência. Se não especificado, a busca será realizada por classe.

classe

Uma string contendo a classe de processo para a busca de jurisprudência. Se não especificado, a busca será realizada por palavras-chave.

base

Um vetor contendo os tipos de documentos a serem buscados. Pode incluir "acordaos" ou "decisoes".

quantidade

O número de resultados desejados. O padrão é 25.

arquivo

O diretório onde os arquivos serão salvos. O padrão é o diretório atual.

Veja também

Exemplos

if (FALSE) {
# Baixar jurisprudência por palavras-chave
stf_jurisprudencia_download(busca = "direitos humanos", base = "acordaos", quantidade = 10, arquivo = ".")

# Baixar jurisprudência por classe de processo
stf_jurisprudencia_download(classe = "ADPF", base = 'acordaos', quantidade = 10, arquivo = ".")
}